requirements.txt
et-xmlfile==1.1.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
selenium==3.141.0
six==1.16.0
urllib3==1.26.4
